Start adding the string syntax:
The initial objective was to be able to initialize chars arrays with
strings like in c:

chr array[10];
set(array, "hello");

The current implementation also allows to give such strings to functions
(funcall("hello")). This doesn't work completely for now as the size
of the parameter is not taken in count. The resulting behavior is that
the size of the parameter array is determined by the size of the given
string and not by the size specified in the declaration:

fn function(chr array[100]) { ... }
function("hello");

Here the size of `array` is 6 and not 100 as expected.

Note: there is a similar problem when we give array to functions as the
size of the array cannot change:

fn function(chr array[100]) { ... }
chr arr[10];
function(arr);

The size of array is 10 (and this is completely normal).

Problem 1: the main problem here may not be the fact that the size of the
array is not taken in count but more that the current syntax for
parameters arrays is confusing. This may be changed -> we shouldn't specify
the size of parameters arrays since it shouldn't be determined at this
point, the size of the memory depend on the allocated size or we don't
allocate anything for parameters in a function definition.

Problem 2: the size of arrays is used a lot at compile time but this is
technically wrong, especially when we consider the size of parameters
arrays. To fix this it would be better to use python functions to
determine the size of the lists.