Move error messages in ErrorManager
All the errors messages have been moved inside the ErrorManager class.
This way, we can easily find all the errors and warnings messages. The
other advantage is that we don't build error messages inside the parser
(or any utility function) anymore which is a big readability
improvement.

The messages have also been improved as now the important elements such
as variable names or types are printed in a different color.

There is still a big issue with error messages: the file name is not
printed with the message. These will be hard to fix. Adding a new syntax
element in parser would be the easiest solution (a line that would be
added by the preprocessor). After such a fix, the way the messages are
printed should be changed to be more standard (path/to/file:line:column).

The column count needs to fixed as well.