Type check on arithmetic operators
There was previously no type check on arithmetic operators, which means
that we could put a chr instead of a int or flt inside an add statement.
The result was a runtime error since python doesn't cast chars into
integers automatically (with operators).

Some warning of unused variables have also been removed in the ast file.