Fix no entry point error
Previously, the no entry point error was recorded even if the parser
hasn't parsed the entire file. This means that when the parser recorded a
syntax and exited directly, the no entry point error was recorded
automatically even if the main function was present. Now, we don't
recorded this error when the parser exit with failure.