Add file locations
The warnings and errors messages have been changed. Now the file name
and the line are indicated. For now, the file name is handled manually
in the parser, however flex positions can save files names. It would be
nice to use this feature instead of using a parser rule as we do now.